2019-3-24 1，修复退出导致的空指针 2，记录右上角最新 3，查看记录详情 4.充值记录

+++ app/src/main/java/cn/fengwoo/sealsteward/bean/RecordListBean.java
@@ -20,8 +20,8 @@ public class RecordListBean implements Serializable {
     private String address;
     private String flowNumber;
     private String id;
-    private Integer latitude;
-    private Integer longitude;
+    private Double latitude;
+    private Double longitude;
     private String stampTime;
 
     public String getAddress() {
@@ -48,19 +48,19 @@ public class RecordListBean implements Serializable {
         this.id = id;
     }
 
-    public Integer getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Integer getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
